Extract `:flow-services` module from `:configuration-cache`
And move types to internal package.
