+++ platforms/core-configuration/flow-services/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.GradleModuleServices
@@ -1,0 +1,1 @@
+org.gradle.internal.flow.services.FlowServices
